Small fixes found in peer review Signed-off-by: fastbike <github@moorhouse.net.nz>

+++ Source/Prometheus.Samples.pas
@@ -84,7 +84,7 @@ type
     /// <summary>
     ///  The count of all observed processes, used for histogram
     /// </summary>
-    MetricCount: Double;
+    MetricCount: Int64;
   end;
 
 { PMetricSamples }
